Fixed the bug where tiles were spawned before the music started

+++ Assets/MagicTiles3/Scripts/SO/Tile_SO.cs
@@ -10,7 +10,7 @@ public enum TileType
 public class Tile_SO : ScriptableObject
 {
     public TileType TileType;
-    public int StepIndex;
+    public float StepIndex;
     public float LongLength;
     public int RowIndex;
 }
